feat: make `no_std` and add `alloc` feature`
